build(wasm): bump beni to the 0.3 series
The 0.3 release takes Array::entry's index as isize and returns nil
for indices beyond the archive's mrb_int width, so the six entry call
sites drop their per-crate mrb_int / i32 casts for the one isize form.

+++ wasm/kobako-io/src/io.rs
@@ -212,7 +212,7 @@ fn puts_one(mrb: &Mrb, self_: Value, val: Value) {
         let ary = unsafe { beni::Array::from_value_unchecked(val) };
         let len = collection_len(mrb, val);
         for i in 0..len {
-            puts_one(mrb, self_, ary.entry(i as i32));
+            puts_one(mrb, self_, ary.entry(i as isize));
         }
         return;
     }
